refactor step "I have multiple categories"
