+++ e2e/src/domains/category/application/use-cases/category-use-cases.factory.ts
@@ -8,6 +8,7 @@ import { SetupCategoryManagementUseCase } from "./setup-category-management.use-
 import { SetupExistingCategoryUseCase } from "./setup-existing-category.use-case";
 import { SetupMultipleCategoriesUseCase } from "./setup-multiple-categories.use-case";
 import { SetupBothCategoryTypesUseCase } from "./setup-both-category-types.use-case";
+import { SetupMultipleTestCategoriesUseCase } from "./setup-multiple-test-categories.use-case";
 import { CreateCategoryWithIconUseCase } from "./create-category-with-icon.use-case";
 import { NavigateToApplicationUseCase } from "./navigate-to-application.use-case";
 import { SearchCategoriesUseCase } from "./search-categories.use-case";
@@ -68,6 +69,12 @@ export class CategoryUseCasesFactory {
     if (!this.categoryService)
       throw new Error("CategoryService required for this use case");
     return new SetupBothCategoryTypesUseCase(this.categoryService);
+  }
+
+  createSetupMultipleTestCategoriesUseCase(): SetupMultipleTestCategoriesUseCase {
+    if (!this.categoryService)
+      throw new Error("CategoryService required for this use case");
+    return new SetupMultipleTestCategoriesUseCase(this.categoryService);
   }
 
   createNavigateToApplicationUseCase(): NavigateToApplicationUseCase {
